MSBG-99:
* moved to utils
